requirements.txt
fastapi
uvicorn
python-dotenv
elevenlabs
deepgram-sdk
openai
httpx
requests